Factor elements in `pip_utils.{hpp,cpp}`
