+++ libmamba/src/api/update.cpp
@@ -4,17 +4,9 @@
 //
 // The full license is in the file LICENSE, distributed with this software.
 
-#include <fmt/color.h>
-#include <fmt/format.h>
-#include <fmt/ostream.h>
-#include <fmt/ranges.h>
-#include <reproc++/run.hpp>
-#include <reproc/reproc.h>
-
 #include "mamba/api/channel_loader.hpp"
 #include "mamba/api/configuration.hpp"
 #include "mamba/api/update.hpp"
-#include "mamba/core/activation.hpp"
 #include "mamba/core/channel_context.hpp"
 #include "mamba/core/context.hpp"
 #include "mamba/core/package_database_loader.hpp"
@@ -25,137 +17,13 @@
 #include "mamba/solver/libsolv/solver.hpp"
 #include "mamba/solver/request.hpp"
 
+#include "pip_utils.hpp"
+
 namespace mamba
 {
     namespace
     {
         using command_args = std::vector<std::string>;
-
-        tl::expected<command_args, std::runtime_error> get_other_pkg_mgr_update_instructions(
-            const std::string& name,
-            const std::string& target_prefix,
-            const fs::u8path& spec_file
-        )
-        {
-            const auto get_python_path = [&]
-            { return util::which_in("python", get_path_dirs(target_prefix)).string(); };
-
-            const std::unordered_map<std::string, command_args> other_pkg_mgr_update_instructions{
-                { "pip",
-                  { get_python_path(), "-m", "pip", "install", "-U", "-r", spec_file, "--no-input" } },
-                { "pip --no-deps",
-                  { get_python_path(), "-m", "pip", "install", "-U", "--no-deps", "-r", spec_file, "--no-input" } }
-            };
-
-            auto found_it = other_pkg_mgr_update_instructions.find(name);
-            if (found_it != other_pkg_mgr_update_instructions.end())
-            {
-                return found_it->second;
-            }
-            else
-            {
-                return tl::unexpected(std::runtime_error(
-                    fmt::format("no update instruction found for package manager '{}'", name)
-                ));
-            }
-        }
-
-        bool reproc_killed(int status)
-        {
-            return status == REPROC_SIGKILL;
-        }
-
-        bool reproc_terminated(int status)
-        {
-            return status == REPROC_SIGTERM;
-        }
-
-        void assert_reproc_success(const reproc::options& options, int status, std::error_code ec)
-        {
-            bool killed_not_an_err = (options.stop.first.action == reproc::stop::kill)
-                                     || (options.stop.second.action == reproc::stop::kill)
-                                     || (options.stop.third.action == reproc::stop::kill);
-
-            bool terminated_not_an_err = (options.stop.first.action == reproc::stop::terminate)
-                                         || (options.stop.second.action == reproc::stop::terminate)
-                                         || (options.stop.third.action == reproc::stop::terminate);
-
-            if (ec || (!killed_not_an_err && reproc_killed(status))
-                || (!terminated_not_an_err && reproc_terminated(status)))
-            {
-                if (ec)
-                {
-                    LOG_ERROR << "Subprocess call failed: " << ec.message();
-                }
-                else if (reproc_killed(status))
-                {
-                    LOG_ERROR << "Subprocess call failed (killed)";
-                }
-                else
-                {
-                    LOG_ERROR << "Subprocess call failed (terminated)";
-                }
-                throw std::runtime_error("Subprocess call failed. Aborting.");
-            }
-        }
-
-        auto update_for_other_pkgmgr(const Context& ctx, const detail::other_pkg_mgr_spec& other_spec)
-        {
-            const auto& pkg_mgr = other_spec.pkg_mgr;
-            const auto& deps = other_spec.deps;
-            const auto& cwd = other_spec.cwd;
-
-            TemporaryFile specs("mambaf", "", cwd);
-            {
-                std::ofstream specs_f = open_ofstream(specs.path());
-                for (auto& d : deps)
-                {
-                    specs_f << d.c_str() << '\n';
-                }
-            }
-
-            command_args update_instructions = [&]
-            {
-                const auto maybe_instructions = get_other_pkg_mgr_update_instructions(
-                    pkg_mgr,
-                    ctx.prefix_params.target_prefix.string(),
-                    specs.path()
-                );
-                if (maybe_instructions)
-                {
-                    return maybe_instructions.value();
-                }
-                else
-                {
-                    throw maybe_instructions.error();
-                }
-            }();
-
-            auto [wrapped_command, tmpfile] = prepare_wrapped_call(
-                ctx,
-                ctx.prefix_params.target_prefix,
-                update_instructions
-            );
-
-            reproc::options options;
-            options.redirect.parent = true;
-            options.working_directory = cwd.c_str();
-
-            Console::stream() << fmt::format(
-                ctx.graphics_params.palette.external,
-                "\nUpdating {} packages: {}",
-                pkg_mgr,
-                fmt::join(deps, ", ")
-            );
-            fmt::print(LOG_INFO, "Calling: {}", fmt::join(update_instructions, " "));
-
-            auto [status, ec] = reproc::run(wrapped_command, options);
-            assert_reproc_success(options, status, ec);
-            if (status != 0)
-            {
-                throw std::runtime_error("pip failed to update packages");
-            }
-        }
 
         auto create_update_request(
             PrefixData& prefix_data,
@@ -370,7 +238,7 @@ namespace mamba
         for (auto other_spec :
              config.at("others_pkg_mgrs_specs").value<std::vector<detail::other_pkg_mgr_spec>>())
         {
-            update_for_other_pkgmgr(ctx, other_spec);
+            install_for_other_pkgmgr(ctx, other_spec, /* update= */ true);
         }
     }
 }
